Remove loop checking whether to fetch more images
While going through images for a given hashtag, there was a check
to see whether we needed to fetch more images for that tag. Since
I'm working with tags that are not popular enough to have fresh
images all of the time, this resulted in a lot of image already
liked warnings. Taking this out as a temporary fix.