Move on to next take after too many already liked
If the user had already liked several consecutive images in a tag,
the bot would keep going through the images and trying to like images
that the user already liked.

Now we set a config variable for max tolerance of consecutive images
already liked. If the bot exceeds that tolerance, we break and move
on to the next tag.